+++ Adapters/BidonAdapterMetaAudienceNetwork/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 6.21.1.0
+* Updated to FBAudienceNetwork 6.21.1
+
 ## 6.21.0.0
 * Updated to FBAudienceNetwork 6.21.0